Bug 1062406 - Part 1 - Change x and y parameters of window.scroll* CSSOM-View DOM calls from double to unrestricted double. r=bz - WebIDL updated so that x and y parameters of window.scroll, window.scrollTo,   and window.ScrollBy are changed from "double" to "unrestricted double". - Implemented mozilla::ToZeroIfNonfinite - Updated nsGlobalWindow::Scroll, ScrollTo, and ScrollBy methods so that they   replace non-finite numbers with 0.

+++ dom/base/nsGlobalWindow.h
@@ -897,11 +897,11 @@ public:
                 mozilla::ErrorResult& aError);
   void ResizeBy(int32_t aWidthDif, int32_t aHeightDif,
                 mozilla::ErrorResult& aError);
-  void Scroll(int32_t aXScroll, int32_t aYScroll,
+  void Scroll(double aXScroll, double aYScroll,
               const mozilla::dom::ScrollOptions& aOptions);
-  void ScrollTo(int32_t aXScroll, int32_t aYScroll,
+  void ScrollTo(double aXScroll, double aYScroll,
                 const mozilla::dom::ScrollOptions& aOptions);
-  void ScrollBy(int32_t aXScrollDif, int32_t aYScrollDif,
+  void ScrollBy(double aXScrollDif, double aYScrollDif,
                 const mozilla::dom::ScrollOptions& aOptions);
   void ScrollByLines(int32_t numLines,
                      const mozilla::dom::ScrollOptions& aOptions);
